SQLQuery / Deserialization progress. Can now handle query of properties of a Document.
Rather than a whole document (select make from Cars).  The serialization format of the
record is different for these two cases.  Both return an ODocument, consistent with the
Java client.  Still only printing out a "ResultSet" which hasn't been fleshed out how
to return such an entity to the user.