First pass at implementing Serialize - in context of EMBEDDEDMAP for SQLCommand params.
Create OEmbeddedMap type, as a map that retains insertion order is needed and it is
much easier to do a type switch off this type.
Added serializationVersion to the DBClient, but it still needs to be updated based
on what the server returns.
Added UNKNOWN type to datatypes in oschema.field