Solved the "mixed header" problem in binserde.
In some cases, OrientDB network binary protocol returns a header of mixed types -
some have ids (to be looked up in the cached GlobalProperties) and others have
names (string) and types (byte), like a property query, even for a document query.
I changed the header datastructure that binserde.go uses for reading OrientDB
headers to handle that and be able to tell which dataPtr goes with which id or
name+type.