Support for serializing DATE and DATETIME from Go structs is now in place.
Added tests for serializing DATE and DATETIME to client.go functional test.
Fixed bug in readDataValue/readDateTime where it was setting time.Unix nanos
(2nd arg) as millis.

Added new ErrDataTypeMismatch error to oerror package - for cases where the user
specifies as type with ODocument#FieldWithType, but the value passed in is
not supported (does not match).  Test added for this to client.go.