Added dbc param to ORecordSerializer#Serialize interface.
The Serialize function needs to look into the globalProperties
map of the database it is serializing for.  The currDB field
of the dbClient changes over time, so this cannot be fixed
in the Serializer, but the serializer needs access.  So decided
to put a field in the ORecordSerializerV0 struct that holds the
reference to the dbc (and thus dbc.currDB.globalProperties)
for the duration of the Serialize or Deserialize call only.

Also, cannot just pass in the globalProperties because in some
instances the ORecordSerializerV0 needs to refresh properties
via a round trip to the database, so it needs to change the
state of the current dbc.currDB.