Removed dirty flag from oschema.ODocument.
This is hard-to-impossible to track and the only value I see
(currently) is that it can prevent a user from doing an update
client-side if nothing has changed.  This is of very minor
benefit for a lot of client driver complexity, so it is removed.
Example of difficult situation: the user can modify []*oschema.OLink
or linkmaps (which are map[string]*oschema.OLink) without going
through an ogonori API.  To truly monitor dirty, you'd have to
disallow all mods except through an ogonori API.

Another issue: the UpdateRecord code would have to iterate
through all Embedded objects and set their dirty flags to
false after a successful update.  This is a potential performance
hit for little value.